Split large multi-env tests into smaller tests
