+++ .github/workflows/tests.yml
@@ -46,11 +46,14 @@ jobs:
           # - test_data_collection  # TODO: https://github.com/StanfordVL/BEHAVIOR-1K/issues/2016
           - test_dump_load_states
           - test_envs
-          - test_multiple_envs_dummy
+          - test_multiple_envs_dummy_api
+          - test_multiple_envs_dummy_scene
           - test_multiple_envs_grasp
           - test_multiple_envs_point_navigation
           - test_multiple_envs_point_reaching
-          - test_multiple_envs_behavior
+          - test_multiple_envs_behavior_infra_api
+          - test_multiple_envs_behavior_infra_scene
+          - test_multiple_envs_behavior_task
           - test_object_removal
           - test_object_states
           # - test_primitives
